backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
pydantic==2.11.5
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.18
openai==1.82.1
pypdf
sqlmodel==0.0.24
alembic==1.16.1
psycopg[binary]==3.2.3
python-multipart==0.0.20
greenlet
jiter>=0.2.0
pgvector
httpx
orjson
python-dotenv
pydantic-settings
